+++ app/src/main/kotlin/com/maltaisn/notes/ui/labels/adapter/LabelListDiffCallback.kt
@@ -23,5 +23,7 @@ class LabelListDiffCallback : DiffUtil.ItemCallback<LabelListItem>() {
     override fun areItemsTheSame(old: LabelListItem, new: LabelListItem) = old.id == new.id
 
     override fun areContentsTheSame(old: LabelListItem, new: LabelListItem) =
-        old.label.name == new.label.name && old.checked == new.checked
+        old.label.name == new.label.name &&
+                old.label.hidden == new.label.hidden &&
+                old.checked == new.checked
 }