+++ app/src/main/kotlin/com/maltaisn/notes/ui/labels/adapter/LabelListViewHolder.kt
@@ -16,6 +16,7 @@
 
 package com.maltaisn.notes.ui.labels.adapter
 
+import androidx.core.view.isInvisible
 import androidx.recyclerview.widget.RecyclerView
 import com.maltaisn.notes.sync.R
 import com.maltaisn.notes.sync.databinding.ItemLabelBinding
@@ -24,6 +25,7 @@ class LabelListViewHolder(val binding: ItemLabelBinding) : RecyclerView.ViewHold
 
     fun bind(item: LabelListItem, adapter: LabelAdapter) {
         binding.labelTxv.text = item.label.name
+        binding.hiddenImv.isInvisible = !item.label.hidden
 
         val view = binding.root