+++ app/src/main/kotlin/com/maltaisn/notes/ui/main/MainActivity.kt
@@ -26,6 +26,7 @@ import androidx.drawerlayout.widget.DrawerLayout
 import androidx.navigation.NavController
 import androidx.navigation.NavDestination
 import androidx.navigation.fragment.NavHostFragment
+//import com.github.venom.Venom
 import com.maltaisn.notes.App
 import com.maltaisn.notes.TAG
 import com.maltaisn.notes.model.converter.NoteTypeConverter